+++ core/src/commonMain/kotlin/com/episode6/mxo2/Builder.kt
@@ -1,8 +1,0 @@
-package com.episode6.mxo2
-
-import com.episode6.mxo2.internal.MockspressoBuilderContainer
-
-/**
- * Root entry point to create a new [MockspressoBuilder]
- */
-fun MockspressoBuilder(): MockspressoBuilder = MockspressoBuilderContainer()